integration test cases: moved upperCase to own dir
