+++ tests/integration-test-cases/upper-case/index.js
@@ -1,0 +1,9 @@
+const onObject = require("./on-object");
+const onCombined = require("./on-combined");
+const onFieldDefinition = require("./on-field-definition");
+
+module.exports = {
+  onObject,
+  onCombined,
+  onFieldDefinition,
+};
