+++ tests/integration-test-cases/upper-case/on-combined.js
@@ -1,5 +1,5 @@
 const { makeExecutableSchema } = require("graphql-tools");
-const { upperCase } = require("./directives");
+const { upperCase } = require("./directive");
 
 const typeDefs = `
   directive @upperCase on FIELD_DEFINITION | OBJECT
